Updated documentation for component extensions
Made a big update on recycling logic, and the method signature has changed a bit. 

Highlight: The build method now has an additional argument where it passes in a `UIView` object if it's being recycled. The developer must check if it's nil. If it's nil, need to instantiate. If not, they can just typecast it to whatever UIView subclass they want and instantiate it and finally return it. 

The commit: https://github.com/Jasonette/JASONETTE-iOS/commit/fff1273ffa08f275787c9f01e2ef3e7940278915